# Add benchmark for undo_index.