--- a/benchmark/CMakeLists.txt
+++ b/benchmark/CMakeLists.txt
@@ -1,0 +1,3 @@
+file(GLOB UNIT_TESTS "bench.cpp")
+add_executable( chainbase_bench bench.cpp  )
+target_link_libraries( chainbase_bench  chainbase ${PLATFORM_SPECIFIC_LIBS} )