This PR adds a `benchmark` directory containing a benchmark designed to exercise the `undo_index` specifically.
I ran it 9 times with the older version of chainbase (before the memory optimization) and identically 9 times with the newer version (using 42 bit offsets as pointers for memory efficiency).

The memory optimized version does not appear to be significantly slower (test with gcc_12, release build, AMD 7950x). These numbers show an average 3.05% slowdown with the memory optimization with intensive chainbase use, so I would not expect to see a noticable slowdown in leap.

| Before memory opt | |
|---------|--------|
|Bench time |     13.39s|
|Bench time |     12.42s|
|Bench time |     12.58s|
|Bench time |     12.36s|
|Bench time |     13.49s|
|Bench time |     12.69s|
|Bench time |     12.41s|
|Bench time |     12.33s|
|Bench time |     12.54s|

| After memory opt   | |
| --------|---------|
|Bench time |     12.93s|
|Bench time |     12.66s|
|Bench time |     12.91s|
|Bench time |     12.76s|
|Bench time |     13.13s|
|Bench time |     12.56s|
|Bench time |     12.54s|
|Bench time |     13.34s|
|Bench time |     12.56s|